ml-model-deployment/ml_model_env/python/deloitte-tensorflow/pyproject.toml
[tool.poetry]
name = "deloitte-tensorflow"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "deloitte_tensorflow"}]

[tool.poetry.dependencies]
python = "~3.8"
tensorflow = "^2.11.0"
numpy = "^1.24.2"
requests = "^2.28.2"
pillow = "^9.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
